Keep the code in spec_shm.h as close as possible to the original while suppressing a potential problem in a comparison between a long and an int.
git-svn-id: https://svn.code.sf.net/p/pymca/code@564 a8702ac3-0063-49a2-b157-4a214673c9c1

+++ PyMca/sps/Src/sps.c
@@ -663,7 +663,7 @@ static int init_ShmIDs (void)
 	== (struct shm_header *) -1)
       continue;
     
-    if (shm->head.head.magic != SHM_MAGIC) {
+    if (shm->head.head.magic != ((long) SHM_MAGIC)) {
       c_shmdt((void *)shm);
       continue;
     }
